+++ CHANGELOG.md
@@ -1,7 +1,28 @@
+## Version 1.0.2
+
+#### Bug fix
+
+- Fix bad annotation checking for the container.
+- Fix some documentation issue again.
+- Handles errors for the AbstractCommand.
+- Fix CommandTestCase bugs related to the usage of Helpers withing the command.
+- Fix not mocked services during the fixture loading.
+- Fix Fixtures creation when the property is inherited.
+
+
+#### New features
+
+- Add handy functions for the fixture creation.
+- Add the `DATE_FORMAT` function to the SQLite language that behave like MySQL function.
+- Add better `ConnectionFactory`.
+- Add the `buildObject` function for all tests.
+
+
+
 ## Version 1.0.1
 
 #### Bug fix
 
-- Fix composer annotation autoloading
-- Fix inspector errors and code cleaning
-- Fix some documentation issue
+- Fix composer annotation autoloading.
+- Fix inspector errors and code cleaning.
+- Fix some documentation issue.